requirements/dev.txt
-r base.txt
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
